+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Bug #20: Update `HasStrokeDashArray` trait and tests to support float values for `stroke-dasharray` attribute (@terabytesoftw)
 - Enh #21: Add `HasOpacity` trait and corresponding tests for managing SVG `opacity` attribute (@terabytesoftw)
 - Bug #22: Add float type support to `stroke-dasharray` attribute and update documentation (@terabytesoftw)
+- Enh #23: Add `HasTransform` trait and corresponding tests for managing SVG `transform` attribute (@terabytesoftw)
 
 ## 0.2.0 March 31, 2024